requirements.txt
numpy==2.2.5
pandas==2.3.3
scikit-image==0.25.2
opencv-python==4.12.0.88
torch==2.9.1
torchvision==0.24.1
imageio==2.37.2
pytorch-transformers==1.2.0
torch-optim==0.0.4
pytorch-nn-tools==0.3.7
pillow==12.0.0
matplotlib==3.10.7
tqdm==4.67.1

